fix sonar
