+++ .scannerwork/report-task.txt
@@ -2,5 +2,5 @@ projectKey=MyImplementationsInReact
 serverUrl=https://sonar.sajermann.com
 serverVersion=9.8.0.63668
 dashboardUrl=https://sonar.sajermann.com/dashboard?id=MyImplementationsInReact
-ceTaskId=AYwldK25JWvWS0rd44Og
-ceTaskUrl=https://sonar.sajermann.com/api/ce/task?id=AYwldK25JWvWS0rd44Og
+ceTaskId=AYwmEZhVJWvWS0rd44Ol
+ceTaskUrl=https://sonar.sajermann.com/api/ce/task?id=AYwmEZhVJWvWS0rd44Ol
