+++ sonar-project.properties
@@ -5,7 +5,7 @@ sonar.sourceEncoding=UTF-8
 # sonar.host.url=http://localhost:9000
 sonar.host.url=https://sonar.sajermann.com
 
-sonar.exclusions=node_modules/**
+sonar.exclusions=node_modules/**, src/Hooks/UseToast/index-*.tsx, src/Hooks/UseCountdown/index.ts
 sonar.coverage.exclusions=test/**/*
 sonar.cpd.exclusions=test/**/*
 sonar.javascript.lcov.reportPaths=coverage/lcov.info
