+++ src/Components/Input/Types/TInput/index.ts
@@ -1,0 +1,11 @@
+import { DetailedHTMLProps, InputHTMLAttributes } from 'react';
+import { TBeforeChange } from '../TBeforeChange';
+
+export type TInput = DetailedHTMLProps<
+	InputHTMLAttributes<HTMLInputElement>,
+	HTMLInputElement
+> & {
+	iserror?: boolean;
+	onBeforeChange?: TBeforeChange;
+	debounce?: number;
+};
