+++ src/Components/Input/index.tsx
@@ -1,34 +1,7 @@
-import {
-	ChangeEvent,
-	DetailedHTMLProps,
-	forwardRef,
-	InputHTMLAttributes,
-	useEffect,
-	useState,
-} from 'react';
+import { forwardRef, useEffect, useState } from 'react';
 import { tv } from 'tailwind-variants';
-import { TCep } from '~/Types/TCep';
-import { TCnpj } from '~/Types/TCnpj';
-import { TCpf } from '~/Types/TCpf';
-import { TCurrency } from '~/Types/TCurrency';
-import { mask } from '~/Utils/Mask';
-
-type TInput = DetailedHTMLProps<
-	InputHTMLAttributes<HTMLInputElement>,
-	HTMLInputElement
-> & {
-	iserror?: boolean;
-	onBeforeChange?: {
-		removeNumber?: boolean;
-		removeUpperCase?: boolean;
-		removeLowerCase?: boolean;
-		removeSpecialCharacter?: boolean;
-		regexForReplace?: RegExp;
-		fn?: (e: ChangeEvent<HTMLInputElement>) => ChangeEvent<HTMLInputElement>;
-		applyMask?: TCurrency | TCnpj | TCpf | TCep;
-	};
-	debounce?: number;
-};
+import { TInput } from './Types/TInput';
+import { onChangeCustom, preOnChange } from './Utils';
 
 const input = tv({
 	slots: {
@@ -69,83 +42,14 @@ export const Input = forwardRef<HTMLInputElement, TInput>(
 			color: iserror ? 'error' : 'primary',
 		});
 
-		async function onChangeCustom(e: React.ChangeEvent<HTMLInputElement>) {
-			if (!onBeforeChange && onChange) {
-				onChange(e);
-				return;
-			}
-
-			const temp = { ...e };
-			let valueTemp = temp.target.value;
-
-			if (onBeforeChange?.removeLowerCase) {
-				valueTemp = valueTemp.replace(/[a-z]/g, '');
-			}
-
-			if (onBeforeChange?.removeUpperCase) {
-				valueTemp = valueTemp.replace(/[A-Z]/g, '');
-			}
-
-			if (onBeforeChange?.removeNumber) {
-				valueTemp = valueTemp.replace(/\d/g, '');
-			}
-
-			if (onBeforeChange?.removeSpecialCharacter) {
-				valueTemp = valueTemp.replace(
-					/[!@#$%&*(),.?":{ }|<>'¨_=+[;^~´`°\]\\\-/]/g,
-					''
-				);
-			}
-
-			if (onBeforeChange?.regexForReplace) {
-				valueTemp = valueTemp.replace(onBeforeChange?.regexForReplace, '');
-			}
-
-			if ((onBeforeChange?.applyMask as TCurrency)?.currency) {
-				valueTemp = mask.real({
-					value: valueTemp,
-					decimalPlace: (onBeforeChange?.applyMask as TCurrency).currency
-						?.decimalPlace,
-				});
-			}
-
-			if ((onBeforeChange?.applyMask as TCnpj)?.cnpj) {
-				valueTemp = mask.cnpj(valueTemp);
-			}
-
-			if ((onBeforeChange?.applyMask as TCpf)?.cpf) {
-				valueTemp = mask.cpf(valueTemp);
-			}
-
-			if ((onBeforeChange?.applyMask as TCep)?.cep) {
-				valueTemp = mask.cep(valueTemp);
-			}
-
-			temp.target.value = valueTemp;
-
-			if (onBeforeChange?.fn && onChange) {
-				const newEvent = onBeforeChange?.fn(temp);
-				onChange(newEvent);
-				return;
-			}
-
-			if (onChange) {
-				onChange(temp);
-			}
-		}
-
-		async function preOnChange(e: React.ChangeEvent<HTMLInputElement>) {
-			if (!debounce) {
-				onChangeCustom(e);
-				return;
-			}
-			setEvent(e);
-		}
-
 		useEffect(() => {
 			const timer = setTimeout(() => {
 				if (debounce && event) {
-					onChangeCustom(event);
+					onChangeCustom({
+						e: event,
+						onBeforeChange,
+						onChange,
+					});
 				}
 			}, debounce);
 
@@ -159,7 +63,15 @@ export const Input = forwardRef<HTMLInputElement, TInput>(
 				className={inputPropsInternal({
 					class: className,
 				})}
-				onChange={preOnChange}
+				onChange={e =>
+					preOnChange({
+						e,
+						setEvent,
+						debounce,
+						onBeforeChange,
+						onChange,
+					})
+				}
 			/>
 		);
 	}
